Unison.Util.Map

Description

Map utilities.

Documentation

bimap :: Ord a' => (a -> a') -> (b -> b') -> Map a b -> Map a' b'

bitraverse :: (Applicative f, Ord a') => (a -> f a') -> (b -> f b') -> Map a b -> f (Map a' b')

bitraversed :: (Ord a', Ord k') => Traversal k k' a a' -> Traversal v v' a a' -> Traversal (Map k v) (Map k' v') a a'

deleteLookup :: Ord k => k -> Map k v -> (Maybe v, Map k v)

Like delete, but returns the value as well.

foldMapM :: (Ord k, Monad m, Foldable t) => (a -> m (k, v)) -> t a -> m (Map k v)

Construct a map from a foldable container by mapping each element to monadic action that returns a key and a value.

The map is constructed from the left: if two elements map to the same key, the second will overwrite the first.

unionWithM :: forall m k a. (Monad m, Ord k) => (a -> a -> m a) -> Map k a -> Map k a -> m (Map k a)

remap :: Ord k1 => ((k0, v0) -> (k1, v1)) -> Map k0 v0 -> Map k1 v1

Reconstruct a map entirely, given a function from old keyvalue to new keyvalue.

remap f = Map.fromList . map f . Map.toList

traverseKeys :: (Applicative f, Ord k') => (k -> f k') -> Map k v -> f (Map k' v)

traverseKeysWith :: (Applicative f, Ord k') => (v -> v -> v) -> (k -> f k') -> Map k v -> f (Map k' v)

swap :: Ord b => Map a b -> Map b a

swap throws away data if the input contains duplicate values

upsert :: Ord k => (Maybe v -> v) -> k -> Map k v -> Map k v

Upsert an element into a map.

upsertF :: (Functor f, Ord k) => (Maybe v -> f v) -> k -> Map k v -> f (Map k v)

Upsert an element into a map.

valuesVector :: Map k v -> Vector v