Unison.Util.Set

Documentation

asSingleton :: Set a -> Maybe a

Get the only member of a set, iff it's a singleton.

difference1 :: Ord a => Set a -> Set a -> Maybe (Set a)

Set difference, but return Nothing if the difference is empty.

differenceMap :: Ord k => Set k -> Map k a -> Set k

Like difference, but the second argument is a map.

foldCommutativeM :: Monad m => (a -> b -> m b) -> b -> Set a -> m b

Fold a set strictly with a monadic "commutative" combining function that doesn't receive the elements in any particular order.

insertMaybe :: Ord a => Maybe a -> Set a -> Set a

intersects :: Ord a => Set a -> Set a -> Bool

Get whether two sets intersect.

mapMaybe :: Ord b => (a -> Maybe b) -> Set a -> Set b

symmetricDifference :: Ord a => Set a -> Set a -> Set a

traverse :: (Applicative f, Ord b) => (a -> f b) -> Set a -> f (Set b)

for :: (Ord b, Applicative f) => Set a -> (a -> f b) -> f (Set b)

flatMap :: Ord b => (a -> Set b) -> Set a -> Set b

filterM :: (Ord a, Monad m) => (a -> m Bool) -> Set a -> m (Set a)

forMaybe :: (Ord b, Applicative f) => Set a -> (a -> f (Maybe b)) -> f (Set b)

thenInsert :: Ord a => Set a -> a -> Set a

thenInsertMaybe :: Ord a => Set a -> Maybe a -> Set a

unalignWith :: (Ord b, Ord c) => (a -> These b c) -> Set a -> (Set b, Set c)

unzipWith :: (Ord b, Ord c) => (a -> (b, c)) -> Set a -> (Set b, Set c)