Unison.Util.Map

Description

Map utilities.

Documentation

bimap :: Ord a' => (a -> a') -> (b -> b') -> Map a b -> Map a' b'

bitraverse :: (Applicative f, Ord a') => (a -> f a') -> (b -> f b') -> Map a b -> f (Map a' b')

deleteLookup :: Ord k => k -> Map k v -> (Maybe v, Map k v)

Like delete, but returns the value as well.

foldMapM :: (Ord k, Monad m, Foldable t) => (a -> m (k, v)) -> t a -> m (Map k v)

Construct a map from a foldable container by mapping each element to monadic action that returns a key and a value.

The map is constructed from the left: if two elements map to the same key, the second will overwrite the first.

unionWithM :: forall m k a. (Monad m, Ord k) => (a -> a -> m a) -> Map k a -> Map k a -> m (Map k a)