Unison.Util.Set

Documentation

asSingleton :: Set a -> Maybe a

Get the only member of a set, iff it's a singleton.

difference1 :: Ord a => Set a -> Set a -> Maybe (Set a)

Set difference, but return Nothing if the difference is empty.

mapMaybe :: Ord b => (a -> Maybe b) -> Set a -> Set b

symmetricDifference :: Ord a => Set a -> Set a -> Set a

traverse :: (Applicative f, Ord b) => (a -> f b) -> Set a -> f (Set b)

flatMap :: Ord b => (a -> Set b) -> Set a -> Set b

filterM :: (Ord a, Monad m) => (a -> m Bool) -> Set a -> m (Set a)

forMaybe :: (Ord b, Applicative f) => Set a -> (a -> f (Maybe b)) -> f (Set b)